backend/requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
deprecation==2.1.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
gradio==4.28.3
gradio_client==0.16.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lancedb==0.6.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.4
ollama==0.1.8
orjson==3.10.1
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pillow==10.3.0
py==1.11.0
pyarrow==15.0.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pylance==0.10.9
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pyngrok==7.1.6
pyparsing==3.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
ratelimiter==1.2.0.post0
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
retry==0.9.2
rich==13.7.1
rpds-py==0.18.0
ruff==0.4.2
semantic-version==2.10.0
semver==3.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
wikiscraper==1.1.9
